+++ jsconfig.json
@@ -2,7 +2,6 @@
   "compilerOptions": {
     "paths": {
       "@/*": ["./src/*"],
-      "@components/*": ["./components/*"]
     }
   }
 }
