+++ src/components/commons/AddBookGroup/AddBookGroup.module.scss
@@ -1,5 +1,5 @@
-@import '@/app/styles/mixins.scss';
-@import '@/app/styles/colors.scss';
+@import '@/styles/mixins.scss';
+@import '@/styles/colors.scss';
 
 .addBtn {
   position: absolute;
