+++ src/components/commons/Book/Book.module.scss
@@ -1,5 +1,5 @@
-@import '@/app/styles/colors.scss';
-@import '@/app/styles/mixins.scss';
+@import '@/styles/colors.scss';
+@import '@/styles/mixins.scss';
 
 .root {
   display: flex;
