+++ src/components/commons/BookPreview/index.js
