+++ src/components/commons/Button/Button.module.scss
@@ -1,4 +1,4 @@
-@import '@/app/styles/mixins.scss';
+@import '@/styles/mixins.scss';
 
 .root {
   img {
