+++ src/components/commons/CreateForm/CreateForm.module.scss
@@ -1,5 +1,5 @@
-@import '@/app/styles/colors.scss';
-@import '@/app/styles/mixins.scss';
+@import '@/styles/colors.scss';
+@import '@/styles/mixins.scss';
 
 .root {
   position: relative;
