+++ src/components/commons/Library/index.js
@@ -1,0 +1,1 @@
+export { default } from './Library';
