+++ src/components/commons/Modal/Modal.module.scss
@@ -1,4 +1,4 @@
-@import '@/app/styles/colors.scss';
+@import '@/styles/colors.scss';
 
 .root {
   position: fixed;
