+++ src/pages/page.module.scss
@@ -1,4 +1,4 @@
-@import './styles/mixins.scss';
+@import '@/styles/mixins.scss';
 
 .root {
   position: relative;
