+++ src/styles/globals.scss
